Scroll problem
So, it turns out that if you give a "document.body.style.overflowY = 'scroll'" it overrides the initial overflow="hidden" of a class "nav-is-open", so you can scroll the page, while menu is opened. 

I just commented out this, cause I don't know is there a way around this, so for now, it's better to stay off.